requirements.txt
python-vlc==3.0.18122
yt-dlp==2023.12.30
requests==2.31.0 